Video_models/Density_models/CGNet/requirements.txt
albumentations==1.3.0
easydict==1.10
geomloss==0.2.6
head==1.0.0
matplotlib==3.7.1
mmcv==2.0.0
numpy==1.24.3
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.4.0
SciencePlots==2.1.0
scikit_learn==1.2.2
scipy==1.12.0
termcolor==2.4.0
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
utils==1.0.2
tensorboard
pandas
pycocotools